requirements.txt
Django==1.8.13
django-haystack==2.4.1
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-markdown2==0.3.0
django-pagedown==0.1.1
django-pure-pagination==0.3.0
django-uuslug==1.1.7
Markdown==2.6.6
markdown2==2.3.1
oEmbed==0.1.2
Pillow==3.3.1
pysolr==3.5.0
python-slugify==1.2.0
pytz==2016.6.1
requests==2.11.1
Unidecode==0.4.19
